requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
opencv-python==4.8.1.78
ultralytics==8.0.196
minio==7.2.0
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-storages==1.14.2
boto3==1.34.0 